backend/pyproject.toml
[project]
name = "react-fastapi-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Gabriel Martin del Campo y Eternod", email = "gmartindelcye@gmail.com"},
]
dependencies = [
    "fastapi>=0.111.0",
    "sqlmodel>=0.0.19",
    "uvicorn>=0.30.1",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.2",
    "pwdlib[argon2]>=0.2.0",
    "python-jose>=3.3.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
